+++ js/components/counter/counterAnimation.js
@@ -1,0 +1,24 @@
+function counterAnimation() {
+
+const counters = document.querySelectorAll('.counter');
+const speed = 200;
+
+counters.forEach(counter => {
+    const updateCount = () => {
+        const target = +counter.getAttribute('data-target');
+        const count = +counter.innerText;
+
+        const inc = target / speed;
+
+        if(count < target) {
+            counter.innerText = count + inc;
+            setTimeout(updateCount, 1);
+        } else {
+            count.innerText = target;
+        }
+    }
+    updateCount();
+})
+}
+
+export { counterAnimation }
