+++ js/main-home.js
@@ -9,6 +9,7 @@ import { aboutCardsData } from './data/aboutCardsData.js';
 import { aboutCardsRender } from './components/aboutCardsRender/aboutCardsRender.js';
 /* about video */
 /* counter */
+import { counterOnScroll } from "./components/counter/counterOnScroll.js";
 /* services */
 /* projects */
 /* news */
@@ -32,26 +33,8 @@ aboutCardsRender('.testclass1', '.testclass2', aboutCardsData);
 
 /* about video */
 /* counter */
-const counters = document.querySelectorAll('.counter');
-const speed = 200;
-
-counters.forEach(counter => {
-    const updateCount = () => {
-        const target = +counter.getAttribute('data-target');
-        const count = +counter.innerText;
-
-        const inc = target / speed;
-
-        if(count < target) {
-            counter.innerText = count + inc;
-            setTimeout(updateCount, 1);
-        } else {
-            count.innerText = target;
-        }
-    }
-    updateCount();
-});
-
+// counterAnimation();
+counterOnScroll();
 /* services */
 /* projects */
 /* news */
